requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-ollama
python-dotenv